Move lists related code to components fodler
