+++ src/server/components/lists/index.ts
@@ -10,3 +10,5 @@ export {
 } from "./lists";
 
 export { listsRoutes } from "./routes";
+export { getServiceLabel, createListSearchBaseLink } from "./helpers"
+export * from "./types";
