+++ src/server/routes/index.ts
@@ -1,6 +1,6 @@
 import { Express } from "express";
 import healthCheckRouter from "./health-check";
-import listsFinderRouter from "./lists";
+import { listsRouter } from "../components/lists/router";
 import developmentRouter from "./development";
 import { sitemapRouter } from "server/components/sitemap";
 import dashboardRouter from "./dashboard";
@@ -8,7 +8,7 @@ import { cookiesRouter } from "server/components/cookies";
 
 export const configureRouter = (server: Express): void => {
   server.use(healthCheckRouter);
-  server.use(listsFinderRouter);
+  server.use(listsRouter);
   server.use(sitemapRouter);
   server.use(developmentRouter);
   server.use(dashboardRouter);
